made code review changes and refactored integration tests into separate package.
