+++ src/main/java/ama/account/UserService.java
@@ -1,10 +1,7 @@
 package ama.account;
 
-import ama.post.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -24,7 +21,7 @@ public class UserService implements UserDetailsService {
     }
 
     @Override
-    public User loadUserByUsername(String username) throws NullPointerException {
+    public User loadUserByUsername(String username) {
         return userRepository.findByUsername(username);
     }
 }
