+++ src/main/java/tests/integration/AuthenticationTests.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.integration;
 
 import ama.*;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestBuilders.formLogin;
