+++ src/main/java/tests/integration/RegistrationTests.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.integration;
 
 import ama.Application;
 import ama.account.User;
