+++ src/main/java/tests/integration/SubmissionPostTests.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.integration;
 
 import ama.*;
